Assignment-01/requirements.txt
requests==2.31.0
pillow==10.2.0
Requests==2.31.0
